+++ test/api_cache_test.dart
@@ -26,35 +26,41 @@ void runApiCacheTests() {
     });
 
     setUp(() {
-      cacheHelper = ApiCacheHelperImplementation.instanceFor(apiCacheManager: mockApiCacheManager);
+      cacheHelper = ApiCacheHelperImplementation.instanceFor(
+          apiCacheManager: mockApiCacheManager);
     });
     String url = "test_url";
     String data = "Test data";
     String cacheKey = "api_cache_$url";
 
     test("setCacheData - Set cache data and return true on success", () async {
-      when(() => mockApiCacheManager.addCacheData(any())).thenAnswer((invocation) async => true);
+      when(() => mockApiCacheManager.addCacheData(any()))
+          .thenAnswer((invocation) async => true);
       final result = await cacheHelper.setCacheData(url, data);
       expect(result, true);
     });
 
     test("getCacheData -  Test caching and retriving data", () async {
       final expectedData = APICacheDBModel(key: url, syncData: data);
-      when(() => mockApiCacheManager.addCacheData(expectedData)).thenAnswer((invocation) async => true);
-      when(() => mockApiCacheManager.getCacheData(cacheKey)).thenAnswer((invocation) async => expectedData);
+      when(() => mockApiCacheManager.addCacheData(expectedData))
+          .thenAnswer((invocation) async => true);
+      when(() => mockApiCacheManager.getCacheData(cacheKey))
+          .thenAnswer((invocation) async => expectedData);
       final result = await cacheHelper.getCacheData(url);
       expect(result, expectedData);
     });
 
     test("isCacheExist - Test cache existence", () async {
-      when(() => mockApiCacheManager.isAPICacheKeyExist(cacheKey)).thenAnswer((invocation) async => true);
+      when(() => mockApiCacheManager.isAPICacheKeyExist(cacheKey))
+          .thenAnswer((invocation) async => true);
 
       bool cacheExistsAfter = await cacheHelper.isCacheExist(url);
       expect(cacheExistsAfter, true);
     });
 
     test("clearCache - Test cache clearing", () async {
-      when(() => mockApiCacheManager.deleteCache(cacheKey)).thenAnswer((invocation) async => true);
+      when(() => mockApiCacheManager.deleteCache(cacheKey))
+          .thenAnswer((invocation) async => true);
 
       var result = await cacheHelper.clearCache(url);
 
@@ -62,7 +68,8 @@ void runApiCacheTests() {
     });
 
     test('clearAllCache - Test clearing all cache', () async {
-      when(() => mockApiCacheManager.emptyCache()).thenAnswer((invocation) async {});
+      when(() => mockApiCacheManager.emptyCache())
+          .thenAnswer((invocation) async {});
 
       // Clear all cache
       await cacheHelper.clearAllCache();